requirements.txt
# requirements.txt
fastapi
websockets
uvicorn
silero-vad
numpy
faster-whisper
ffmpeg-python
torch
torchaudio